ui: rename insightsApi.ts to txnInsightsApi.ts
This rename is in preparation for splitting the insights
api call into separate txn and stmt calls.

Epic: none

Release note: None
