+++ pkg/ui/workspaces/cluster-ui/src/selectors/insightsCommon.selectors.ts
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-import { FlattenedStmtInsights } from "src/api/insightsApi";
+import { FlattenedStmtInsights } from "src/api/txnInsightsApi";
 import {
   mergeTxnInsightDetails,
   flattenTxnInsightsToStmts,
